Time: 3 ms (34.39%), Space: 8.9 MB (81.27%) - LeetHub

+++ 3412-permutation-difference-between-two-strings/3412-permutation-difference-between-two-strings.cpp
@@ -5,9 +5,9 @@ public:
         size_t tLen = t.length();
         int answer = 0;
 
-        for (size_t i = 0; i < sLen; ++i)
+        for (int i = 0; i < sLen; ++i)
         {
-            for (size_t j = 0; j < tLen; ++j)
+            for (int j = 0; j < tLen; ++j)
             {
                 if (s[i] == t[j])
                 {
